@@ -1,42 +1,44 @@
 package com.customproject.coffeeshop.api.config
 
-import com.fasterxml.jackson.annotation.JsonInclude
-import com.fasterxml.jackson.databind.DeserializationFeature
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.SerializationFeature
-import com.fasterxml.jackson.databind.module.SimpleModule
-import com.fasterxml.jackson.datatype.hibernate5.Hibernate5Module
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
+//import com.fasterxml.jackson.databind.DeserializationFeature
+//import com.fasterxml.jackson.databind.ObjectMapper
+//import com.fasterxml.jackson.databind.SerializationFeature
+//import com.fasterxml.jackson.databind.module.SimpleModule
+//import com.fasterxml.jackson.datatype.hibernate5.Hibernate5Module
+//import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import com.zaxxer.hikari.HikariConfig
 import com.zaxxer.hikari.HikariDataSource
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.context.annotation.Bean
+//import org.springframework.boot.context.properties.ConfigurationProperties
+//import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Primary
+//import org.springframework.context.annotation.Primary
 import javax.sql.DataSource
 
 
 @Configuration
 class UserDataSourceConfig {
-    @Bean
-    @ConfigurationProperties(prefix = "user.datasource.hikari")
+// uncomment on use
+//    @Bean
+//    @ConfigurationProperties(prefix = "user.datasource.hikari")
     fun hikariConfig(): HikariConfig {
         return HikariConfig()
     }
 
-    @Bean
-    @Primary
+// uncomment on use
+//    @Bean
+//    @Primary
     fun dataSource(): DataSource {
         return HikariDataSource(hikariConfig())
     }
 
-    @get:Bean
-    val mapper: ObjectMapper = ObjectMapper().apply {
-        configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
-        configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false)
-        disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
-        registerModule(JavaTimeModule())
-        registerModule(SimpleModule())
-        registerModule(Hibernate5Module())
-    }
+// uncomment on use
+//    @get:Bean
+//    val mapper: ObjectMapper = ObjectMapper().apply {
+//        configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+//        configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false)
+//        disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
+//        registerModule(JavaTimeModule())
+//        registerModule(SimpleModule())
+//        registerModule(Hibernate5Module())
+//    }
 }
